digital_petition/requirements.txt
streamlit
pycryptodome
pandas

